Changed to searching from first available square with fewest options
Changed the solve algorithm so that when it starts guessing numbers it
starts from the square with the fewest options. This decreases the time
it takes to find a solution. It was setup like that before but I forgot
why I did it that way and changed it to just be the first square.
Changed save code to output numbers in lined up columns, although it
still needs work, it's not pefect yet. Using string.format would be a
better way to do it.