Refactoring in the main form
Changed form logic to load Board objects instead of arrays of Strings.
Loading arrays decoupled the form from the Board class, but that
required an intermediate step to render a Board and code to convert a
Board to an array, and just overall more lines of code for no benefit.
Also added a toString() method to the Square class to allow easy
conversion to string.
Converted a lot of unchanging ints to constants just to make it clear
they don't change, and renamed to use the C# constant naming convention
consistently.
Allowed the grid of text boxes to be reused if it's being replaced by a
board that's the same size to save a little render time. I also
consolidated the text box styling code into three methods for the three
styles in use.